\name{runExample}
\alias{runExample}
\title{Run one of the test examples.}
\usage{
  runExample(name = NULL, all = FALSE, thisR = FALSE)
}
\arguments{
  \item{name}{Character name of example.}

  \item{all}{Run all the test examples?}

  \item{thisR}{Run inside this R?}
}
\description{
  Compile and run a test example (\code{runExample()} shows
  all available examples).
}

